fixed remarks and return of status value in print string routine in lab11
